Add internal resolution setting with 6.75x (1080p) option
Wires up the previously-dead multiplier field on the SDL renderer so
players can pick an internal render scale from the Tools menu. 6.75x
is included since it maps GBA's 160px height exactly onto 1080.

+++ app/src/main/cpp/android/sdl/main.h
@@ -51,7 +51,7 @@ struct mSDLRenderer {
 	unsigned height;
 	int viewportWidth;
 	int viewportHeight;
-	int ratio;
+	float ratio;
 
 	bool lockAspectRatio;
 	bool lockIntegerScaling;
